AF-258 Add no security test properties.
