+++ user-microservice/src/main/java/com/aline/usermicroservice/config/WebSecurityConfig.java
@@ -2,15 +2,16 @@ package com.aline.usermicroservice.config;
 
 import com.aline.core.annotations.WebSecurityConfiguration;
 import com.aline.core.dto.response.UserResponse;
-import com.aline.core.model.user.User;
 import com.aline.core.model.user.UserRole;
 import com.aline.core.security.config.AbstractWebSecurityConfig;
 import com.aline.core.security.service.AbstractAuthorizationService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 
 @WebSecurityConfiguration
+@Slf4j(topic = "Web Security Configuration")
 public class WebSecurityConfig extends AbstractWebSecurityConfig {
 
     @Override
@@ -22,15 +23,15 @@ public class WebSecurityConfig extends AbstractWebSecurityConfig {
                         "/users/otp-authentication",
                         "/users/password-reset-otp")
                 .permitAll()
-                .antMatchers(HttpMethod.PUT, "/users/password-reset");
+                .antMatchers(HttpMethod.PUT, "/users/password-reset").permitAll();
     }
 
     @Bean
     public AbstractAuthorizationService<UserResponse> authService() {
         return new AbstractAuthorizationService<UserResponse>() {
             @Override
-            public boolean canAccess(UserResponse responseBody) {
-                return (responseBody.getUsername().equals(getUsername()) ||
+            public boolean canAccess(UserResponse returnObject) {
+                return (returnObject.getUsername().equals(getUsername()) ||
                         getRole() == UserRole.ADMINISTRATOR ||
                         getRole() == UserRole.EMPLOYEE);
             }
