+++ user-microservice/src/test/resources/application-no-security-test.properties
@@ -1,0 +1,1 @@
+app.security.disable-web-security=true
